Refactor and update content configuration
Refactored content configuration by merging the content and config files and introducing image loading through the glob loader. Updated packages to improve compatibility and performance, including an upgrade to the latest AstroJS and Tailwind versions, and replaced non-standard image component usage in the Carousel component with a native HTML img tag. These changes enhance the maintainability and functionality of the project.